+++ src/main/java/com/mraof/minestuck/data/MinestuckEnUsLanguageProvider.java
@@ -2534,6 +2534,8 @@ public class MinestuckEnUsLanguageProvider extends MinestuckLanguageProvider
 		addSubtitles("horn", "Horn honks");
 		addSubtitles("screech", "Estrogen-empowered screech");
 		addSubtitles("upcheladder", "Echeladder levels up");
+		addSubtitles("captchalogue_item", "Updated card in Sylladex");
+		addSubtitles("captchalogue_shuffle", "Updated whole Sylladex");
 		addSubtitles("boondollar_get", "Received Boondollars");
 		addSubtitles("electric_shock", "Electric Shock");
 		addSubtitles("electric_autoharp_stroke", "Chord plays on electric autoharp");