+++ src/main/java/com/mraof/minestuck/inventory/captchalogue/QueueModus.java
@@ -3,8 +3,10 @@ package com.mraof.minestuck.inventory.captchalogue;
 import com.mraof.minestuck.item.CaptchaCardItem;
 import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.item.components.MSItemComponents;
+import com.mraof.minestuck.util.MSSoundEvents;
 import net.minecraft.core.NonNullList;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.fml.LogicalSide;
 
@@ -36,6 +38,7 @@ public class QueueModus extends StackModus
 		
 		if(id == CaptchaDeckHandler.EMPTY_SYLLADEX)
 		{
+			player.level().playSound(null, player.getX(), player.getY(), player.getZ(), MSSoundEvents.EVENT_CAPTCHALOGUE_SHUFFLE.get(), SoundSource.AMBIENT, 1F, 1F);
 			for(ItemStack item : list)
 				CaptchaDeckHandler.ejectAnyItem(player, item);
 			list.clear();